[ImportVerilog] Add a new constructor for vistor and fix building problem
